% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{calc_inclination}
\alias{calc_inclination}
\title{Calculate segment inclination angle}
\usage{
calc_inclination(segments)
}
\arguments{
\item{segments}{An sf object with columns \verb{pX, pY, pZ, kX, kY, kZ}.}
}
\value{
The same sf object with added columns \verb{deltaX, deltaY, deltaZ},
and \code{inclination_angle}.
}
\description{
Angle of the segment relative to the horizontal plane, in degrees.
0 = horizontal, 90 = vertical (downward).
}
