% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_metrics.R
\name{calc_root_length}
\alias{calc_root_length}
\title{Calculate total length per root}
\usage{
calc_root_length(df)
}
\arguments{
\item{df}{A data.frame with columns \code{ROOT_ID} and \code{length_3d}.}
}
\value{
The same data.frame with added column \code{total_length_root}.
}
\description{
Sums segment lengths within each ROOT_ID.
}
