% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{calc_segment_azimuth}
\alias{calc_segment_azimuth}
\title{Calculate segment azimuth}
\usage{
calc_segment_azimuth(segments)
}
\arguments{
\item{segments}{An sf object with columns \verb{pX, pY, kX, kY}.}
}
\value{
The same sf object with added column \code{azimuth}.
}
\description{
Azimuth (bearing) from start to end point, measured clockwise from north
in degrees (0-360).
}
