% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{calc_segment_length}
\alias{calc_segment_length}
\title{Calculate segment 3D length}
\usage{
calc_segment_length(segments)
}
\arguments{
\item{segments}{An sf object with columns \verb{pX, pY, pZ, kX, kY, kZ}.}
}
\value{
The same sf object with added column \code{length_3d}.
}
\description{
Euclidean distance between start (pX,pY,pZ) and end (kX,kY,kZ) of each segment.
}
