% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{Load and Prepare Root Data}
\description{
Functions for loading raw GPR root CSV data exported from
software like ImpulseRadar, REFLEXW, or GPR-SLICE and preparing
it for spatial analysis.
}
