% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{points_to_sf}
\alias{points_to_sf}
\title{Convert root points to sf object}
\usage{
points_to_sf(df, crs = 2178)
}
\arguments{
\item{df}{A data.frame with X, Y, Z columns.}

\item{crs}{EPSG code or CRS object. Default \code{2178}.}
}
\value{
An sf point object with XYZ geometry.
}
\description{
Converts a data.frame with X, Y, Z columns to an sf point object with
XYZ dimensions.
}
