% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{split_survey_column}
\alias{split_survey_column}
\title{Split survey column}
\usage{
split_survey_column(
  df,
  survey_col = "Survey",
  sep = "_",
  plot_position = 4,
  plot_name = "PLOT_ID",
  drop_original = TRUE
)
}
\arguments{
\item{df}{A data.frame.}

\item{survey_col}{Name of the column to split. Default \code{"Survey"}.}

\item{sep}{Separator string. Default \code{"_"}.}

\item{plot_position}{Which part (after split) contains the plot ID. Default \code{4}.}

\item{plot_name}{Name for the new plot column. Default \code{"PLOT_ID"}.}

\item{drop_original}{Drop the original survey column? Default \code{TRUE}.}
}
\value{
A data.frame with a new plot column.
}
\description{
Splits a composite survey/filename column into parts by a separator
and extracts the plot identifier.
}
